+++ src/handlers/index.ts
@@ -1,5 +1,6 @@
 export * from './simple';
 export * from './complete';
+export * from './descendants';
 export * from './project';
 export * from './pick';
 export * from './path';